requirements.txt
amqp (1.0.11)
anyjson (0.3.3)
argparse (1.2.1)
async (0.6.1)
billiard (2.7.3.28)
boto (2.9.5)
celery (3.0.19)
distribute (0.6.34)
Fabric (1.6.1)
gitdb (0.5.4)
GitPython (0.3.2.RC1)
Jinja2 (2.7)
kombu (2.5.10)
MarkupSafe (0.18)
MySQL-python (1.2.4)
paramiko (1.10.1)
pycrypto (2.6)
python-dateutil (2.1)
pytz (2013b)
PyYAML (3.10)
six (1.3.0)
smmap (0.8.2)
SQLAlchemy (0.8.1)
wsgiref (0.1.2)
